resume_parser/requirements.txt
streamlit
pdfplumber
python-docx
requests
python-dotenv
pymysql
sentence_transformers 
langchain_community
faiss-cpu